+++ experitur/__init__.py
@@ -1,6 +1,5 @@
-from experitur.core.context import Context, experiment, push_context, run
-from experitur.core.samplers import GridSampler, MultiSampler, RandomSampler
-from experitur.core.trial import TrialProxy
+from experitur.core.experiment import Experiment
+from experitur.core.trial import TrialParameters
 from experitur.errors import ExperiturError
 
 from ._version import get_versions