+++ tests/test_dox_py.py
@@ -3,8 +3,8 @@ import os
 
 import pytest
 
-from experitur.core.context import Context, push_context
-from experitur.dox import load_dox, DOXError
+from experitur.core.context import Context
+from experitur.dox import DOXError, load_dox
 
 
 @pytest.fixture(name="dox_py_fn")
@@ -14,10 +14,10 @@ def fixture_dox_py_fn(tmp_path):
         f.write(
             inspect.cleandoc(
                 """
-                from experitur import experiment
+                from experitur import Experiment
 
-                @experiment(
-                    parameter_grid={
+                @Experiment(
+                    parameters={
                         "a1": [1],
                         "a2": [2],
                         "b": [1, 2],
@@ -27,7 +27,7 @@ def fixture_dox_py_fn(tmp_path):
                     return trial.parameters
 
                 # This experiment shouldn't be executed, because this combination of callable and parameters was already executed.
-                experiment(
+                Experiment(
                     "second_experiment",
                     parent=baseline
                 )
@@ -42,7 +42,7 @@ def test_dox_py(dox_py_fn):
     wdir = os.path.splitext(dox_py_fn)[0]
     os.makedirs(wdir, exist_ok=True)
 
-    with push_context(Context(wdir)) as ctx:
+    with Context(wdir) as ctx:
         load_dox(dox_py_fn)
 
         # Execute experiments
@@ -64,8 +64,7 @@ def test_unknown_extension(unknown_fn):
     wdir = os.path.splitext(unknown_fn)[0]
     os.makedirs(wdir, exist_ok=True)
 
-    with push_context(Context(wdir)) as ctx:
-
+    with Context(wdir):
         with pytest.raises(DOXError):
             load_dox(unknown_fn)
 
@@ -83,7 +82,6 @@ def test_malformed_py(malformed_py_fn):
     wdir = os.path.splitext(malformed_py_fn)[0]
     os.makedirs(wdir, exist_ok=True)
 
-    with push_context(Context(wdir)) as ctx:
-
+    with Context(wdir):
         with pytest.raises(DOXError):
             load_dox(malformed_py_fn)